requirements.txt
apipkg==1.5
attrs==20.2.0
boto3==1.14.60
botocore==1.17.60
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
docutils==0.15.2
execnet==1.7.1
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
jmespath==0.10.0
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.2
pytest-dotenv==0.5.2
pytest-forked==1.3.0
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-testconfig==0.2.0
pytest-xdist==2.1.0
python-dateutil==2.8.1
python-dotenv==0.14.0
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
toml==0.10.1
urllib3==1.25.10
zipp==3.1.0
jsonschema==3.2.0
jsonref==0.2
selenium
beautifulsoup4==4.8.1